pyproject.toml
[project]
name = "vintagewisdom"
version = "0.1.0"
description = "Personal decision support system based on historical cases"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.7",
  "pyyaml>=6.0",
]

[project.scripts]
vw = "vintagewisdom.cli.commands:main"

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "ruff>=0.5",
  "mypy>=1.8",
]

kg = [
  "neo4j>=5.0",
  "redis>=5.0",
]

ingest = [
  "pypdf>=4.2",
  "python-docx>=1.1",
]

web = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "python-multipart>=0.0.9",
]

graphrag = [
  "qdrant-client>=1.9",
  "httpx>=0.27",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.10"
warn_unused_ignores = true
warn_return_any = true
strict_optional = true
